Test gateway persistence
