+++ tests/test_gateway.py
@@ -1,8 +1,11 @@
 """Test the gateway."""
+import asyncio
+from unittest.mock import call
+
 import pytest
 
 from aiomysensors.exceptions import InvalidMessageError
-from aiomysensors.gateway import Gateway
+from aiomysensors.gateway import Config, Gateway
 from aiomysensors.model.message import Message
 from aiomysensors.model.protocol import PROTOCOL_VERSIONS
 
@@ -63,3 +66,34 @@ async def test_unset_protocol_version(message, message_schema, node, child, tran
     assert gateway.transport.writes == [
         message_schema.dump(Message(0, 255, 3, 0, 2, ""))
     ]
+
+
+async def test_persistence(mock_file, persistence_data, transport):
+    """Test gateway persistence."""
+    mock_file.read.return_value = persistence_data
+    gateway = Gateway(transport, Config(persistence_file="test_path.json"))
+
+    assert not gateway.nodes
+
+    async with gateway:
+        assert mock_file.read.call_count == 1
+        assert gateway.nodes
+        node = gateway.nodes.get(1)
+        assert node.battery_level == 0
+        assert node.children
+        child = node.children.get(1)
+        assert child.child_id == 1
+        assert child.child_type == 38
+        assert child.values
+        value = child.values.get(49)
+        assert value == "40.741894,-73.989311,12"
+
+        await asyncio.sleep(0.1)
+
+        # First write call done by the save task.
+        assert mock_file.write.call_count == 1
+        assert mock_file.write.call_args == call(persistence_data)
+
+    # Second write call done when stopping save task.
+    assert mock_file.write.call_count == 2
+    assert mock_file.write.call_args == call(persistence_data)
